requirements.txt
langchain>=0.3.26
langchain-core>=0.3.67
langchain-community>=0.3.0
langchain-openai>=0.0.8
langgraph>=0.0.20
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
streamlit>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
faiss-cpu>=1.7.4
pypdf>=3.9.0
python-multipart>=0.0.6
pytest>=7.4.0 